refactor

+++ lib/cache.cc
@@ -141,10 +141,10 @@ void cache_t::update_lru(uint64_t index, uint64_t mru_way) {
 }
 
 void cache_t::stats() {
-   printf("\taccesses   = %lu\n", accesses);
-   printf("\tmisses     = %lu\n", misses);
+   printf("\taccesses   = %llu\n", (unsigned long long)accesses);
+   printf("\tmisses     = %llu\n", (unsigned long long)misses);
    printf("\tmiss ratio = %.2f%%\n", 100.0*((double)misses/(double)accesses));
-   printf("\tpf accesses   = %lu\n", pf_accesses);
-   printf("\tpf misses     = %lu\n", pf_misses);
+   printf("\tpf accesses   = %llu\n", (unsigned long long)pf_accesses);
+   printf("\tpf misses     = %llu\n", (unsigned long long)pf_misses);
    printf("\tpf miss ratio = %.2f%%\n", 100.0*((double)pf_misses/(double)pf_accesses));
 }
